Separate classes into separate files and add tests for pathOf
