+++ format/src/main/scala/org/calinburloiu/music/microtonalist/format/ComposedScaleRepo.scala
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2021 Calin-Andrei Burloiu
+ *
+ *    Licensed under the Apache License, Version 2.0 (the "License");
+ *    you may not use this file except in compliance with the License.
+ *    You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *    Unless required by applicable law or agreed to in writing, software
+ *    distributed under the License is distributed on an "AS IS" BASIS,
+ *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *    See the License for the specific language governing permissions and
+ *    limitations under the License.
+ */
+
+package org.calinburloiu.music.microtonalist.format
+
+import com.google.common.net.MediaType
+import org.calinburloiu.music.intonation.{Interval, Scale}
+
+import java.net.URI
+
+trait ComposedScaleRepo extends ScaleRepo {
+  def getScaleRepo(uri: URI): Option[ScaleRepo]
+
+  override def read(uri: URI): Scale[Interval] =
+    getScaleRepoOrThrow(uri).read(uri)
+
+  override def write(scale: Scale[Interval], uri: URI, mediaType: Option[MediaType]): Unit =
+    getScaleRepoOrThrow(uri).write(scale, uri, mediaType)
+
+  protected def getScaleRepoOrThrow(uri: URI): ScaleRepo = getScaleRepo(uri)
+    .getOrElse(throw new BadScaleRequestException(uri))
+}
