+++ format/src/main/scala/org/calinburloiu/music/microtonalist/format/FileScaleListRepo.scala
@@ -18,26 +18,21 @@ package org.calinburloiu.music.microtonalist.format
 
 import org.calinburloiu.music.microtonalist.core.ScaleList
 
+import java.io.{FileInputStream, FileNotFoundException}
 import java.net.URI
+import scala.util.Try
 
-trait ScaleListRepo {
-  def read(uri: URI): ScaleList
+class FileScaleListRepo(scaleListFormat: ScaleListFormat) extends ScaleListRepo {
+  override def read(uri: URI): ScaleList = {
+    val path = pathOf(uri)
+    val inputStream = Try {
+      new FileInputStream(path.toString)
+    }.recover {
+      case e: FileNotFoundException => throw new ScaleListNotFoundException(uri, e.getCause)
+    }.get
 
-  def write(scaleList: ScaleList, uri: URI): Unit
+    scaleListFormat.read(inputStream, Some(baseUriOf(uri)))
+  }
+
+  override def write(scaleList: ScaleList, uri: URI): Unit = ???
 }
-
-/**
- * Exception thrown if the requested scale list could not be found.
- */
-class ScaleListNotFoundException(uri: URI, cause: Throwable = null)
-  extends RuntimeException(s"A scale list with $uri was not found", cause)
-
-
-/**
- * Exception thrown if the the scale list request was invalid.
- */
-class BadScaleListRequestException(uri: URI, message: Option[String] = None, cause: Throwable = null)
-  extends RuntimeException(message.getOrElse(s"Bad scale list request for $uri"), cause)
-
-class ScaleListReadFailureException(val uri: URI, message: String, cause: Throwable = null)
-  extends RuntimeException(message, cause)
